anaimation for colours & file reorg
