+++ tracer/src/lib/textures/texture-check-solid.ts
@@ -4,7 +4,7 @@
 //
 
 import { Texture } from './texture';
-import { Colour } from './colour';
+import { Colour } from '../colour';
 
 export class TextureCheckSolid extends Texture {
   solid = true;
