+++ tracer/src/raytracer.ts
@@ -113,8 +113,8 @@ export class Raytracer {
       // Image row complete
       cacheCorner = null;
       bufferY++;
-      const perc: number = Math.round((bufferY / this.task.sliceHeight) * 100);
-      if(bufferY % Math.floor(this.task.sliceHeight / 10) == 0) console.log(`### Frame ${this.task.frame}, task '${this.task.index + 1} / ${this.task.jobId}' rendered ${perc}%`);
+      //const perc: number = Math.round((bufferY / this.task.sliceHeight) * 100);
+      //if(bufferY % Math.floor(this.task.sliceHeight / 10) == 0) console.log(`### Frame ${this.task.frame}, task '${this.task.index + 1} / ${this.task.jobId}' rendered ${perc}%`);
     }
 
     return this.image;
